functions/image-size/requirements.txt
gunicorn==20.1.0
Flask==2.2.5
Werkzeug==3.0.1
opencv-python ==4.9.0.80
Pillow==10.2.0